+++ Cargo.toml
@@ -27,6 +27,7 @@ with-syntex = ["syntex", "pnet_macros/with-syntex"]
 [dependencies]
 ipnetwork = "0.12"
 pnet_base = { path = "pnet_base" }
+pnet_sys = { path = "pnet_sys" }
 pnet_datalink = { path = "pnet_datalink" }
 
 [dependencies.clippy]
