+++ pnet_sys/Cargo.toml
@@ -1,10 +1,17 @@
 [package]
-name = "pnet_base"
+name = "pnet_sys"
 version = "0.19.0"
 authors = ["Robert Clipsham <robert@octarineparrot.com>", "Linus Färnstrand <faern@faern.net>"]
 license = "MIT/Apache-2.0"
 homepage = "https://github.com/libpnet/libpnet"
 repository = "https://github.com/libpnet/libpnet"
-description = "Fundamental base types and code used by pnet."
-keywords = ["networking", "ethernet"]
+description = "Access to network related system function and calls."
+keywords = ["networking", "datalink", "ethernet", "raw"]
 categories = ["network-programming"]
+
+[dependencies]
+libc = "0.2"
+
+[target.'cfg(windows)'.dependencies]
+winapi = "0.2.*"
+ws2_32-sys = "0.2.*"
