+++ pnet_sys/src/lib.rs
@@ -1,4 +1,4 @@
-// Copyright (c) 2014-2016 Robert Clipsham <robert@octarineparrot.com>
+// Copyright (c) 2014, 2015 Robert Clipsham <robert@octarineparrot.com>
 //
 // Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
 // http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
@@ -6,5 +6,5 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-mod macaddr;
-pub use macaddr::*;
+pub mod internal;
+pub mod sockets;
